+++ doc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # What's New with Enterprise Web Components
 
+## 1.0.0
+
+### 1.0.0 Fixes
+
+- `[Header]` Changed the header color setting to work separate from personalization. ([#2050](https://github.com/infor-design/enterprise-wc/issues/2050))
+- `[Personalizations]` Added a method to reset the personalization color to the default. ([#2054](https://github.com/infor-design/enterprise-wc/issues/2054))
+
 ## 1.0.0-beta.23
 
 ### 1.0.0-beta.23 Ui Changes